[push] Golang API boilerplate.
